First release version

+++ listrooms.php
@@ -3,6 +3,14 @@
  <body>
 
 <?php
+
+include "checksession.php";
+include "header.php";
+include "menu.php";
+echo '<div id="site_content">';
+include "sidebar.php";
+echo '<div id="content">';
+
 include "config.php"; //load in any variables
 $DBC = mysqli_connect(SERVERNAME, DBUSER, DBPASSWORD, DBDATABASE);
 
@@ -19,7 +27,12 @@ $result = mysqli_query($DBC,$query);
 $rowcount = mysqli_num_rows($result); 
 ?>
 <h1>Room list</h1>
-<h2><a href='addroom.php'>[Add a room]</a><a href="/bnb/">[Return to main page]</a></h2>
+<h2>
+  <?php  if(isset($_SESSION['loggedin']) && $_SESSION['loggedin']  ==1){?>
+  <a href='addroom.php'>[Add a room]</a>
+  <?php } ?>
+  <a href="/bnb/">[Return to main page]</a>
+</h2>
 <table border="1">
 <thead><tr><th>Room Name</th><th>Type</th><th>Action</th></tr></thead>
 <?php
@@ -30,9 +43,12 @@ if ($rowcount > 0) {
 	  $id = $row['roomID'];	
 	  echo '<tr><td>'.$row['roomname'].'</td><td>'.$row['roomtype'].'</td>';
 	  echo     '<td><a href="viewroom.php?id='.$id.'">[view]</a>';
-	  echo         '<a href="editroom.php?id='.$id.'">[edit]</a>';
+
+    if(isset($_SESSION['loggedin']) && $_SESSION['loggedin'] ==1){
+    echo         '<a href="editroom.php?id='.$id.'">[edit]</a>';
 	  echo         '<a href="deleteroom.php?id='.$id.'">[delete]</a></td>';
-      echo '</tr>'.PHP_EOL;
+    }
+    echo '</tr>'.PHP_EOL;
    }
 } else echo "<h2>No rooms found!</h2>"; //suitable feedback
 
@@ -65,6 +81,10 @@ mysqli_close($DBC); //close the connection once done
         }
         ?>
 
+<?php
+echo '</div></div>';
+include "footer.php";
+?>
 
 </body>
 </html>
